+++ packages/bot/src/common/services/parsers.service.ts
