+++ packages/bot/src/common/standard-behaviors/behaviors/stats.behavior.ts
@@ -1,10 +1,10 @@
-import { ChangeLog } from "@root/platform/discord/rollem-bot/changelog";
-import { Logger } from "@root/platform/discord/rollem-bot/logger";
-import { HandlerType, PromLogger } from "@common/prom-logger";
+import { ChangeLog } from "@common/services/changelog/changelog";
+import { Logger } from "@common/services/logger.service/logger.service";
+import { HandlerType, PromLogger } from "@common/services/prom-logger.service/prom-logger.service";
 import { BehaviorContext } from "@common/standard-behaviors/types/behavior-context";
 import { BehaviorResponse } from "@common/standard-behaviors/types/behavior-response";
 import { BehaviorBase, Trigger } from "@common/standard-behaviors/behavior.base";
-import { BehaviorStatsBase } from "@common/standard-behaviors/stats-base";
+import { BehaviorStatsBase } from "@common/standard-behaviors/stat.behavior.base";
 import { Injectable } from "injection-js";
 import { toPairs } from "lodash";
