+++ packages/bot/src/common/standard-behaviors/stat.behavior.base.ts
