+++ packages/bot/src/common/util/types/utility-types.ts
