+++ packages/bot/src/platform/discord/rollem-bot/behaviors/event-monitor.behavior.ts
@@ -1,9 +1,9 @@
-import { DiscordBehaviorBase } from "./discord-behavior-base";
+import { DiscordBehaviorBase } from "./discord.behavior.base";
 import util, { promisify } from "util";
 import { Client, Message, User, MessageReaction, ClientEvents } from "discord.js";
-import { Logger, LoggerCategory } from "@root/platform/discord/rollem-bot/logger";
+import { Logger, LoggerCategory } from "@common/services/logger.service/logger.service";
 import { Injectable } from "injection-js";
-import { PromLogger } from "@common/prom-logger";
+import { PromLogger } from "@common/services/prom-logger.service/prom-logger.service";
 
 interface AdditionalMetrics {
   guildId?: string;
