+++ packages/bot/src/platform/discord/rollem-bot/discord-config.service.ts
